requirements.txt
streamlit
requests
faiss-cpu
sentence-transformers
PyPDF2
python-dotenv
